Use the same lint config as in the HarpGuru project
This means that I'm not extending from the @react-native-community file
anymore which is definitely not ideal, but I'm going to be copying files
over from HarpGuru directly and it just makes sense for the styling to
not have to be changed.

Hopefully I will be given sufficient assistance from the tests which I
write to spot when something severely illfitting for native has been used.