Have HarpRow handle a single row of the matrix
This is a deviation from the webapp design because native can't do the
same css grid display as web. This seems fine though and I'm going to
architect this version slightly different anyway since there's a good
chance that I'm going to want to keep as much non react stuff away from
the react stuff as possible in order to make it reusable between the web
and native apps.