Publish with src as entry point
This actually seems to be consumed by the expo project successfully
which is good. The reason it works presumably is because it's already
set up to interpret the typescript and esnext format files which the src
folder contains.

I don't know if this is worth fixing or whether building the lib folder
is a waste of time.