Align testing tech in harpstrata with harpguru-core
..somewhat

We don't need to explicitly use babel-jest in harpguru-core, but I think
that's just because it's bundled up in the @testing-library preset.

ts-lint doesn't seem necessary since we have a typechecking step in the
`ci` run script and it's a piece of technology which was easier to throw
away in harpstrata than to add to harpguru-core.

There are also some changes to the formatting of package.json and some
alteration to the available run scripts made in this commit. Not ideal,
but give me a break. I'm being pretty fucking methodical with this whole
workspaces experiment in general.