Add recent changes to changelog
This branch is titled 'experiment-as-npm-package', but really the
intention was to package the HarpGuru app core away from the expo
boilerplate so that it could potentially be run from other contexts.

Recognising that these could sit in a monorepo architecture along with
the harpstrata project, making future development easier, so in the end
the repackaging has been successful, but we've got no npm publishing
configuration in place. No need for the new goal.