Add recent changes to CHANGELOG.md
This project is now just a boilerplate for running the HarpGuru core
code, and whats more, it's able to run it from a yarn workspaces
context.

I'll now try to get these projects working from a lerna monorepo. If
that can be done well then all of these component projects can be
archived.