Initialise lerna project `npx lerna init -i`
I'm unsure about the versioning type to prefer here.

I have gone with independent for a few reasons:

1. the harpstrata project already has npm versions published
and is a long way ahead of harpguru-core if we ever create an
npm package of that.

2. I only have a small number of packages so managing the versions
should not be too dificult.

3. I like having CHANGELOGs in the fomat I have them at the moment, each
of which identifying whether a change has been MAJOR, MINOR or PATCH.