Prefer to start lock on pozition
The logic for this is that I think the most common first piece of
knowledge the user will have is either which harp they have or which
pitch they want to play in. Then they will want to lock that in while
they modify the other covariant they know. I think that's more common
than wanting to start with modifying the position.

Anyway, it's kind of arbitrary and at least it's symmetrical by being in
the middle.