Improve scaling variable calculation
Now we can modify the menuHiddenScale value and expect the label
protrusion to stay in tact. It doesn't remain as the original input
value, because the entire object gets scaled, but it does at least stay
consistent with itself.

This means that if we need to add another menu, we simply need to change
the menuHiddenScale value to 0.33333 and each menu label tab will fill a
third of the height of the screen