Replace harp cell dedicated hook with common util
As described in the previous commit, this functionality is required in
two different components now so it's been made common.

I'm not entirely happy with the name or the fact that you need to
provide both degree and pitch ids regardless of which one you're
presenting but these can be improved in due course.