Move the display mode option to layout menu
It's still a tricky one, but with the inclusion of the activity legend,
there is less need for the user to switch between display modes while
they're chosing their covariables.

Having the covariant menu include the display mode definitely makes it
feel cluttered, so I'll try this for a while and see how it feels.