Remove unused styles and compromise animation fail
The observation about the icons not being effected by the scaling
translations in the previous commit were incorrect. What they *are*
uneffected by is the opacity translations.

To compromise I have just had the icons shrink away as the menu expands.
This means that they disappear and don't distract during the animation
without disappearing so abrumptly that they are distracting for a
different reason.