Use RenderedTone component in quiz presentation
The just means that we present the quiz question in the same form as the
tones wich are displayed in the harp face and activity legend.

The implementation in this commit is a bit of a hack. I don't think the
way that I've made the RenderedTone component adaptive to this new
context. This needs reworking.

I've also noticed that sometimes two tones are presented rather than
one. F# Gb for example. However in other questions, just the F# might be
presented. Something weird going on there.