Add the close button to the layout menu
This is required for consistency, but at the moment it does seem a
little strange to have this button here when a tap anywhere on the menu
will close it.

I think the tap handler should be restricted to only apply to the open
and close icons.

FYI the `tapHandler` function both opens and closes the menu depending
on the current state.