Rearrange option vertical layout for safety
There was not enough control over the distance from the edges that the
elements were set at which meant that there was a good chance that the
lock icon would end up somewhere unattractive with respect to the title.

There was also a chance that the options list would overlap with the
title if we make it any longer.