Improve tap handler callback function name
The name now gives a clue about what it's doing. It's a duel purpose
function, but the icons it's attached to are only available during one
or other of its functions so they look like they are consistently either
an open of a close button, but really they aren't. If you could  access
the close button while the menu is already closed then it would open the
menu.