Lay foundations for reduced global state
The plan is to use the wrapper to derive all of the facts relevant to
the harpcell before passing them in as props, so that we can memoise
components. The momoisation has not been added yet, but it will be once
this strategy is proven working.

There is at least one potential flaw here which is the inclusion of the
toggle function. That toggle function uses global state and might
therefore evoke a rerender automatically when the state changes. I'm not
sure, I'm sure it will be possible to work around it, but let's see.