Move hook and callback out of cells
The idea is that everything the cells need is either passed as a static
param or imported as a static resource.

The memoisation is still not working, even with the `areEqual` function
always returning true, we still see a console log if we include one in
the HarpCellAccessible body.

The `getSizes` function is still in use in the stylesheets which is
possibly one problem. We would want to take that out to only be
calculated once, even if it's not the source of this rerendering.

We probably want to do the same with all of the hooks calls which are
currently in the wrapper. We get one wrapper instance for each cell.
Better would be to call these hooks from the harp face fragments since
there's only 3 or 4 of them, and pass the results down as props.