Move global values from cells to rows
This will reduce the number of global value retrievals.

I have not moved the setter as well just in case it doesn't work. I'll
do that on a separate commit so it can be reverted if necessary.

These values can be pulled up even further from the rows to the
fragments.