Improve pan even handler to reduce lag
The state update which is performed in the event handler is a heavy
operation. The fact that it was being performed when the pan handler was
in a BEGAN state was the cause of the slow tap handler state transition.

What was happening was that pan handler and the tap hander would enter a
BEGAN state at the same time because they are both under the finger
press, the pan handler would block on the state update for 500ms and
then the states would advance as it becomes clear that this is a tap
rather than a pan.

The state doesn't need to be updated until we know that we are entering
a state which actually contributes to the conclusion that we have some
pan related action to perform. That's why we can wait until the END
state before we do anything.

We possibly could get away with doing this in the ACTIVE state if we set
the pan properties appropriately to not enter active until a certain
distance had been traced.