Move the global state calls back in to harp cell
It turns out that the majority of the slowdown was caused by the faulty
tap handler which was enhanced in
https://github.com/js-jslog/harpguru/pull/28

Consequentially, I think we've learned that there's not much to be
gained from pulling the global state calls further up the chain, if
anything at all.

The main advantage of this architecture now is that we can memoise the
harp cells below the tap handler and potentially do some of the
following:

1. Handle the BEGIN state of the tap handler seperately from the END
state and evoke a rerender of the child cell with some kind of colour
shift, to mark the start of the interaction. This colour shift will need
to be rerendered back out again if the tap handler is cancelled.

2. An animation value could be added to each cell so achieve something
more fluid than what's described in 1. above. This was tried before, but
it made the harpface extremely unresponsive due to the amount of time
each cell took to rerender. Given that we will typically only be
rerendering 4 cells at a time rather than 70 due to the memoisation,
this might now be a realistic thing to try again.