Simplify cell styles further
Removing styles which default to the required values means there's less
to read, but we need to be careful it's still clear what the intention
is for the styles. I'm not so sure that it's clear that these styles are
invisible. I'll run like this for a while and see how I get on.