Force batch update to prevent *loads* of renders
When a harp cell is touched without this forced batching in place, it
takes a *really* long time to re-render.

There is a gap in the react-native types definition so typescript things
that there is in fact no unstable_batchedUpdates to use. But there is!

I will need to fill that gap somehow or the project won't build.