Remove redundant state and correct used one
The bufferedActivityToggles object should start life empty unless we
want some kind of scale to be immediately presented when the app starts.

This might be quite a good way of forcing the resizing to happen before
the user interacts with the harp-face, helping get the caching ready for
the first actual interaction.