Replace ActiveDegreeIds type where semantically inaccurate
From a type perspective this change has no effect at all. However, the
identical `ReadonlyArray<DegreeIds>` type is more semantically accurate
since all of the replacements here are not actually *the* active degree
ids in the app. Sometimes they are not even in a function which has to
be aware that it is dealing with active degree ids (batchToggleDegreeIds)