Add all possible dependencies to `useEffect`
I don't think that this is necessary. I think that only the
`bufferedActivityToggles` is actually needed. However, I don't think
there is a significant efficiency cost here, and the possibility of
introducing strange bugs is too much of a concern with new
functionality, so I'm going to play it safe and log it as an area to be
investigated during optimisation.