Increase timeout on active harpstrata update
Now that the user has instant feedback on their actual action, they can
afford to wait a bit longer for the full update. This may also be true
in quiz mode, but only if we reset the question timeout after each
input, otherwise they won't have a chance to input all the cells before
the timer runs out.