Return to original options, with minor animation addition
The "new" option style is not working out. This was only ever supposed
to be a stop gap to make sure that there is *some* kind of animation in
place for the user to receive feedback. This implementation will do for
now while I make more video's and make plans for a proper menu
enhancement.