Bring menu label style in to component
The information provided by the get-menu-styles-and-animation-vals
function is looking like it needs to be simplified in to just providing
some absolutely core shared values, while the spefiic values look like
they'll be drawn out in to specialised components like the one we're
making now.

This means that the covariant menu label is now broken, but it'll be
fixed soon once it's also using this component we're developing.