Uncollapse elements to fix bug
The collapsing of the View element in the the Animated.View element
which was done in a previous commit was a mistake. The component was no
longer bounded withing it's container and the buttons were invisibly
overlapping, which meant that only one of them could be pressed.

My analysis of the bug in the previous commit was way off, thankfully. I
thought it didn't make any sense.