Create dedicated MenuFace component
The main benefit of this is just that it's a styled component.

You can see that we've taken our first step in to defining static menu
values as distinct from the dynamic ones. Already we can see that we've
got some confusion outstanding in where we should define the menu
values.