Reset quiz question timer when buffer is updated
This will mean that the user has an opportunity to press all the
locations of the question tone before the answer is given.

The amount of time they have will just be the same length of time that
the buffer usually waits to be flushed though which isn't long. Perhaps
it should be increased in we're in quiz mode.