Add workspace config justification to README.md
Some of the decisions have been made using common sense with an
awareness of my own gaps in knowledge. I'm not convinced that this is
optimally configured, but until I've discovered what's wrong I need to
record my reasons for the decisions so that I can evaluate them
confidently later on.