Include type polyfills for dependants
If the dependants are importing a non-compiled version of the package
then they will need to benefit from the very same type polyfills that
this project needs. Since these files won't be automatically imported
from the canonical file paths in the local project when the project is
an import, we need to make sure they are included by other means.