Add optimised typescript analysis
The idea is that I can run this static analysis along with the limited
tests before any commit is accepted. Unfortunately it seems like even
the incremental typescript analysis will take 30 seconds to complete.
I'll live with this for a bit and see how it goes.

Incremental typescript analysis does not work when Emit=false so we have
had to update typescript to v4 too.