Implement lint-staged for eslint
This is particularly difficult in a monorepo.

Each package may have it's own eslint config, and that config may depend
on packages which are not available from the monorepo root. This means
that lerna must be used to run from each project scope, and that means
that we will be passing the file paths down to each package, 2 of which
will always be missing at least one of the files being staged. This is
why the no-error-on-unmatched-pattern is important.

I removed the json matcher from the lint-staged config so that eslint
itself doesn't have to exclude it, because *that* would lead to a
warning and *that* would mean that the lint-staged would fail.