Upgrade eslint-plugin-react
This was an attempt to work around the need for configuration of the
react version used; currently the lack of configuration is generating a
warning message.

It does say in the docs that in the future the config will have a
default, but right now it doesn't.

Upgrading didn't help in this respect, but it doesn't hurt to have the
latest version if all the linting still passes.