Revert "Add an npm package to temporarily permit time test"
This reverts commit 880c6baa2a67e40a5d63ef8b71872920c47d5e57.

The tests demonstrated that the tested functions execute at
the same speed as before the isActiveMatrix was removed.

There _may_ have been a slight increase in the `getHarpStrata`
function, by a matter of 1 or 2ms though I can't anticipate
why that would be. There's a good chance that it's just
circumstancial.