Add apparatus & prepare homogenising getters
The structure I'm moving towards now will have all of the organised
parts organised in the orderd-parts folder. This looks like a bit of
duplication, but it means that the degree, pitch, pozition etc folders
can really focus on just defining the objects and then the service
functions can be grouped together in the same folder and present a
consistent interface to these objects where they are similar even though
when you look closely they are all very different in content and
generation.