Improve return type of function
We don't expect an undefined value to be returned so we don't want the
type system to expect undefineds as acceptable. Throwing an error is the
right thing to do here short of being able to actually type up the
inbound types so that we would correctly not expect undefined options to
be part of the record object.