Utilise es6 Maps for strict ordering
The Maps aren't perfect in terms of immutability (but we can work around
that by providing non-mutating accessors) and they aren't perfect in
terms of types (they don't require all enums to be entered as keys, and
they don't guarantee a defined response, but we can work around that
with some thrown errors).

Certainly not extatic with it, but it might be the best I can do.