Specify peer dependencies in support packages
I have plans to add multiple support libraries which will all depend on
harpparts, but which will not be connected to one another. I these will
all be required by harpguru-core along with harpparts. It makes sense that
the support libraries make the demand of the peer (harpguru-core) for
harpparts.