Remove unused exports
These exports might have been more useful before when the two covariant
functionalities were part of separate packages, but now that they are
together the number of exported functions can be reduced.

We actually don't need to export any of the covariance-series
functionality here, because it's not used at the moment. It took a lot
of work though and it could well be useful if we make the covariant menu
in harpguru-core more elaborate.