Add simple scales objects as orderable parts
I'm not sure whether these should be orderable parts in the same way as
the others at the moment. The point of the orderable parts is that they
are either fundamentally ordered or are at least going to appear in a
menu in the order arranged here. Scales don't order well. If anything
they 'group' better than they 'order'.

This isn't a massive concern for right now since we only need access to
the scales in order to know what they are rather than for presentational
purposes. We can come back to this question when we actually have a
usecase for ordering vs grouping.