Restrict colour in flash notification to just a moment
This is much more descreet than the previous implementation as well as
being easier to read and also giving the user a sense of snappiness
and reward which is in keeping with the occasions that this type of
animation should be employed.

The actual implementation of this is not ideal. It would be more
efficient if the animations were all triggered upon the displayability
being set to true, and were configured to do their individual arcs from
that point and before the displayability is set to false rather than
being triggered back from the actual setting of the displayability to
false.

This would be clearer and would also result in a more refined animation.
I'll need to brush up on other ways to produce animation values than
just relying on the `useTimingTransition` from the redash library which
I've used successfully for everything else so far.