Use the flash notification wrapper for quiz questions
This brings some consistency to the flash displays.

I've been forced to bring some of the animation elements together since
they weren't behaving well when seperated. It probably makes sense to
try and have all the transformations in one place if possible anyway.

I'm not entirely happy with the syntax or naming of the
additionalScaleFactor, but it will do for this commit.