Wrap the two apps in their own providers
I needed to also bring the cell interaction with the global state in to
the functionality before I was convinced that it was working.

It wouldn't have been necessary to duplicate all of this code actually.
I could have just put harpguru itself in to some providers, but then I
would have had a harder time getting the global object in to each
provider call since I'd have had to create it in App.js rather than just
use the global object which we already create.

I'm sure that I shouldn't keep a global context in the default context
as well as in the providers so There shold rather be a util which
creates the object to be passed in to the providers.