Create a dedicated harp-guru-page component
So now we have harp-guru still at the outside, really just setting up
the two contexts and then handing off to each of the harp-guru-page
instances.

I imagine that the harp-guru object will eventually still require some
of it's own state in order to manage the coordination of which of the
pages to display at the very least. There may be other executive
functions which this object still needs to perform too.