Set global state from a local util
We have some redundancy in the type definition for the global state now.
I have a global.d.ts and the local type definition in the new utils
function. I would like to get rid of one or the other of these, but
first I need to understand which one is correct to have.