Correct the page number labelling
Previously it was showing the *next* page number, rather than *this*
page number.

I've used a map to identify where the next page should be. This would
potentially give us the ability to have multiple maps available,
shortcutting one of the pages if the user only wanted 2 pages for
example.