Deactivate scale notification in quiz mode
It looks like there is some kind of nasty bug in play if the scale
notification happens at the same time as the next quiz question is
flashed. The overlay gets locked on and the app becomes unreachable
behind it.

It would be better to figure out what the cause of the bug is, in case
it reemerges elsewhere, but it seems like this solution is reliable and
I don't have time to investigate right now. I'll need to remember in
case we add more flash notifications which might overlap in the future.