Add hacky colour override for next-page-button
It was just too bold having the two other menu's standing out as the
dark colour like it was. It turns out that the answer is to
differentiate the next page button from the menu's by colour.

This is a very hacky solution to that requirement, but we can afford it.