Add sixth menu tab
This commit has been made without the capacity to check how this looks
visually so the scaling might be off.

I've also not yet actually created the new menu tab to be included so
we'll currently have a duplicate of one of the other menu tabs in the
new position.