Add override as separate state to force flush
This is only a partial implementation of the intended, because I have
realised halfway through that the design is flawed. Calling the
use-flush-buffered-activity-toggles hook a second time means that there
are two separate hook contexts being created. The intention was to give
the quiz cycle some control of the base flusher. To do that I will need
to have the hook called just once instead and pass the returned
functions to the quiz component.