Simplify effect drivers to reduce effect repetition
The introduction of the `bufferedActivityToggles` to this effect were
the cause of the observation made in the TODO comment.

Removing it was the cure, and the plan to address the lost behaviour of
being able to enter solutions for more than the Listening states initial
timeout will be found through alternate means.