Move override flush timings out to quiz component
The problem with the previous approach was that the effects weren't
being re-triggered when a new interval was submitted because it would
just be the same value.

Perhaps using an object instead of a primitive would have been another
solution, but I tihnk it makes good sense for the timings to be
controlled from here too.