Simplify and make more explicit function coverage
It is much more simple to imagine that we simply want to respond to user
toggles immediately if we're not in the listening states. We will mostly
want to reset the harpstrata to being inactive, but to do reliably from
a rendering perspective, we first need to flush so that the toggled
harpcell is ready to be rerendered. If we try to rerender it as blank
immediately then the caching on that object will conclude that since
it's receiving the same props as last render it doesn't need to render
again.