Remove unnecessary flushes and reposition reset
The flushes in the bufferedActivityToggles useEffect should be
sufficient and it's convenient to have them all organised in the same
place.

The resetActiveHarpStrata function call location change might not have
any functional effect, but I think I prefer it here semantically.
Previously it made more sense to say "Whenever we move to the Ask state,
we want to make sure that everything is cleared down", but now we've got
the implicit cleardown following all harpstrata updates made in the
activeHarpStrata useEffect instead.