Use dispatcher to avoid stale variables
When useEffects need to make use of variables which they must not be
driven by the changes of, we need to use functional state changes. In
ReactN this can be achieved with dispatchers to ensure that the latest
version of the state is used, rather than the latest version in the
scope of the useEffect.