Replace toggle buffer flush function with reactn dispatch
For the same reason as a dispatch is preferable for the quiz cycle. We
don't want to have to pass all of the dependency variables in as drivers
of the useEffect, so to make sure that these dependency variables are up
to date when we use them, dispatch can be used.