Split the toggle flushing dispatch updates apart
It seems like trying to update both of the global properties in a single
dispatch might lead to a perceived update to all of the properties in
the global state, while updating individually like this means that we
only get a single update for each one.

It's possible that there's a better way to do this with reactn's
dispatches, but this will do for now. I have not tested whether this
results in two sets of renders and whether we might benefit from
reintroducing unstable_batchUpdates to streamline that.

The mystery posed in the extended comment in this commit is a real one
and one I would like to get to the bottom of.