Create alternate toggle buffer flusher
I'd rather allow the global state to reference the
bufferedActivityToggles and the activeHarpStrata and
update them both in a single command rather than have
to force the batch update like I am currently. I have
a feeling this takes longer to render for some reason
though so I'm going to test them in parallel.