Test the new toggle flusher and add batch update wrapper
Without the batch update wrapper we see a significant delay in
processing when calling it from a `setTimeout` context.

I expect that react is successfully batching these updates automatically
when they are done outside the async context of the `setTimeout` (there
is no delay when it's called from that context at least).

It might be that there is a better way to achieve this or there is
something fundamentally wrong with calling Promise returning functions
in an async function, but this seems to work and I have more
experimentation to do before settling on this approach broadly, so I'll
conduct those investigations before I get bogged down understanding
this.