Remove unnecessary complexity
The function parameter version of this state update is not required. I
put it in place as protection against the callback being run too late
when control should have been given to the other `useEffect` block, but
in practice this seems unnecessary and decluttering this file is a
priority where safe to do so.