Improve comments in quiz hook file
The comments need to be good because this file is quite a mess. I could
split it out in to smaller hooks, but the thing is there's quite a lot
of interdependency, especially with respect to the `resetHarpFace`
function (hense the large comment), which I don't want to easily
overlook when modifying this logic by packaging it up elsewhere. So I
think it's best to leave it all together, even if it's messy. I think
it's just about simple to follow now.