Propose 'use-dimensions' as a replacement solution
The `useWindow` hook option which comes with this alternative library
seems to produce fewer renders. Possibly because it only get's updated
when the window size changes rather than when the screen size changes
too.

This is important at app initialisation when the screen reorients and
the soft menu bar takes effect on the screen shape.