Empty toggle buffer for all but 1st page
I don't want there to be renders happening at the point that the app
loads which aren't seen. It's a waste of renders. If it feels wrong to
not have cells highlighted on the other screens then we can just set
them in the activeHarpStrata rather than in the bufferedActivityToggles.