Create empty scales menu
This menu isn't going to be populated with the same `Option` component
that the other menu's have so I want to start from a semi blank slate.

The plan is to have a standard scrolling menu without the tap animations
that the others have but the selection will be obvious because of the
flash notification which will be presented whenever a new scale is
selected.