Refactor for persistent list
We don't want the list to rerender to the top of the list whenever it's
parent is rerendered.

I was expecting to have to use memo and useCallback here, but it seems
like just making sure that none of the parents variables are accessed
through closure but rather passed down as parameters has been
sufficient.