Add selector arrows to the menu title
The idea is that these menu arrows will give the flexibility to move
through a list of arbitraty length eventually.

I believe that arrows and crosses are safe places to start rather than
swipe actions since they are a more clear call to action.

I learned that lesson the hard way with the very first production
release of the app.