Add scale categorisation in to prepare list split
The UI needs to present the chords and the scales as seperate lists,
even if it's convenient to model them the same way as one another in the
harpparts project.

This might not be the final design, but the complicated bit is where I
want to spend my time right now. I can come back to this once I've
actually got the list working.