Stack scrollable menu lists
Now that they are stacked we can use animation values to make just one
visible at a time.

This is a very ugly implementation, but it took a while to get it
working so I want to save it off before tidying it up. I think the key
was in giving the absolute element some size properties.