Make OptionList more generic - step 1
I don't want to overoptimise or over generalise this too early on, but I
know that I'm going to want to put other lists than just scales in to
this thing so in the course of tidying it up this makes sense to do. It
will make it's intention clearer too.