Memoise stack so animation value not reset each render
We also need to make a decision whether we force the caller to wrap the
tap handler functions in useCallback or we just don't test the tap
handler in the equlity checker.

It would be more secure to force useCallback usage but I don't know how
to make it really clear that this is required. I can predict that the
dispatch function will remain unchanged in effect and will only change
because it's redefined in each render. Practically this is probably ok,
but it's too much of a smell for me to ignore really. I'll probably find
a way to communicate the need for the useCallback.