Correct callback memoisation from previous commit
The previous commit asked the question whether we should ignore the
changing assignment of the tapHandler or not. Well through poor use of
the useCallback function I was inadvertently doing that while thinking
that I wasn't. Without setting the dependencies like I am in this commit
we are opening the door to stale closures if they are a possibility at
all in this situation.