Give option list and title equal gutters
The arrangement of the styles is starting to show it's flaws now since
these gutters aren't actually defined in the same place. We should fix
this.

I also wonder whether I might be making life more difficult for myself
by trying to use flex at the foundations of this. It seems unnecessary
now since the components are defined as being a fixed distance from the
edges.