Add limit check to option stack traversal
This ensures that when you get to the end of the stack you need to start
coming back the other way.

We could have the stack loop, but that presents a number of problems
which need solutions which are probably a lot harder to resolve:

1. How to get the animation value back to 0 without the interpolation
taking it back through the intermediate values.

2. How to remove the arrows when the stack is only one list deep. We
don't want the user to be clicking the arrows thinking that they aren't
working.

3. It seems to be considered that looping lists are potentially
confusing UI anyway.