Animate the pointer visibility in sync with rest
By rest I mean the rest of the option-stack component.

It's important to only use the transition value for the opacity and not
the pointerEvents property because we don't want any delay in making it
disabled if it needs to be. Using the absolute value rather than the
transition is probably safer in this respect (although now I think about
it we can always perform an absolute check on a transition value..)