Refactor code inpreparation for shared use
I do wonder whether the `access-parts` functions should all go inside
the `utils` folder to fit the strover pattern of the react native
packages. I've been a bit flexible in the harpparts package because it's
almost just a repository, but the functionality is growning as we can
see here and there may come a time when the structure needs a good
rethink.