Make test of `itemTapHandler` equality check all optionProps
It's not enough to check just the first optionProps. It could be that
just one of them has been updated, but that means they all unfortunately
need to be rerendered. If that actually produces more rerenders given
the way we're using these components then I'll be surprised and we'll
have to do something about it, but it'll be easier to find out this way
than with stale references.

I should consider just throwing an error if anything in here is found to
not be equal since that would imply something unexpected has happened.